增加人物 把所有数值改成小数 疲劳缓慢修正 修复怪异背包遍历错误 修复移动影响疲劳时报错的bug 修复传送过短距离不会影响疲劳值（例如到车上、到床上） 修复过快移动疲劳值暴涨的情况 疲劳值更加友好

+++ src/HamsterYDS/UntilTheEnd/api/UntilTheEndApi.java
@@ -166,7 +166,7 @@ public class UntilTheEndApi {
         }
 
         public static int getValue(Player player, String type) {
-            return PlayerManager.check(player, type);
+            return (int) PlayerManager.check(player, type);
         }
 
         public static void setValue(Player player, String type, int value) {
